requirements.txt
requests
python-telegram-bot==20.5
httpx
